rename my codpart by functional block
