+++ src/index.html
@@ -44,7 +44,7 @@
           <button class="button-close hide">&#10005</button>
         </div>
         <div class="category">
-          <button class="button">all</button>
+          <button class="button" id="all">all</button>
         </div>
 
         <button class="cart" id="cart">
