+++ src/index.js
@@ -2,6 +2,6 @@ import "./styles/index.scss"; // для компиляции css?
 import "regenerator-runtime/runtime"; // для чего?
 import './Milia' // js Милии
 import './Elena' // js Лены
-import './yulia' // js Юлии
+import './search-block' // js Юлии
 import './cart-button' //скрипт для Корзины
 import './Alena' //прелоадер
