+++ src/styles/_global.scss
@@ -74,7 +74,7 @@ $primaryColor: rgb(45, 155, 146);
 }
 
 //////////////////////////////////////////////////////////
-@import "yulia.css"; //search-block
+@import "search-block.css";
 
 .search__block {
   display: flex;
